requirements.txt
python-dotenv~=1.1.0
injector~=0.22.0
redis~=4.6.0
requests~=2.32.4
Flask~=3.1.1
WTForms~=3.2.1
SQLAlchemy~=2.0.41
pytest~=8.4.1
openai~=1.60.1
langchain~=0.2.17
numpy~=1.26.4
tiktoken~=0.8.0
jieba~=0.42.1
langgraph~=0.2.8
pydantic~=2.10.6
PyYAML~=6.0.2
duckduckgo_search~=8.1.1
torch~=2.7.1
transformers~=4.53.0
celery~=5.4.0
marshmallow~=3.26.0
Werkzeug~=3.1.3
PyJWT~=2.10.1
alembic~=1.16.2